% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imessager.R
\name{send.imessage}
\alias{send.imessage}
\title{Send iMessages from R}
\usage{
send.imessage(recipient, message)
}
\arguments{
\item{recipient}{Recipient phone number or email registered with 'iMessage'? STRING}

\item{message}{What message would you like to send? STRING}
}
\description{
A short and simple function you can use to send 'iMessages' from R running 'macOS' 10.8.x or later. You must have 'Messages.app' configured for the user account you are running 'R' on.
}
\examples{
send.imessage("5551234567","Hello World!")
}
\keyword{notification}
